+++ ecs_push.tf
@@ -24,6 +24,7 @@ data "aws_iam_policy_document" "push_ecs_task_policy" {
       aws_ssm_parameter.db_reader_host.arn,
       aws_ssm_parameter.db_ssl.arn,
       aws_ssm_parameter.default_country_code.arn,
+      aws_ssm_parameter.hsts_max_age.arn,
       aws_ssm_parameter.log_level.arn,
       aws_ssm_parameter.push_host.arn,
       aws_ssm_parameter.push_port.arn,