+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 	- push_ecs_autoscale_scale_down_adjustment
 	- push_ecs_autoscale_scale_up_adjustment
 - Added: SUPPORT_REQUESTS.md
+- Added: Headers Cache-Control, Pragma, Strict-Transport-Security
+- Changed: Push API TLS 1.2 enforcement
 
 
 ## [v0.1.7] 2020-09-17